# public/swagger/openapi.yaml
openapi: 3.0.0
info:
  title: 'Laravel App'
  description: 'Laravel Apis'
  contact:
    name: 'Suraj Sharma'
    email: sharmasuraj41@gmail.com
  version: 1.0.0
servers:
  -
    url: 'http://127.0.0.1:8000'
paths:
...
  /api/login:
    post:
      tags:
        - User
      summary: 'User login'
      description: 'User login'
      operationId: users.Login
      parameters:
        -
          $ref: '#/components/parameters/X-localization'
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - email
                - password
              properties:
                email:
                  type: string
                password:
                  type: password
              type: object
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiModel'
  /api/logout:
    post:
      tags:
        - User
      summary: 'User Logout'
      description: 'User Logout'
      operationId: users.logout
      parameters:
        -
          $ref: '#/components/parameters/X-localization'
      requestBody:
        content:
          multipart/form-data: {  }
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiModel'
      security:
        -
          Bearer: []
  /api/profile_update:
    post:
      tags:
        - User
      summary: 'User profile update'
      description: 'User profile update'
      operationId: users.profileupdate
      parameters:
        -
          $ref: '#/components/parameters/X-localization'
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - name
                - user_name
                - avatar
                - email
                - user_role
              properties:
                name:
                  type: string
                user_name:
                  type: string
                avatar:
                  type: file
                email:
                  type: string
                user_role:
                  type: string
              type: object
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiModel'
      security:
        -
          Bearer: []
components:
  schemas:
    ApiModel:
      required:
        - statusCode
        - msg
        - data
      properties:
        statusCode:
          type: integer
          format: int32
        msg:
          type: string
        data:
          type: object
      type: object
      example:
        statusCode: 200
        message: successful
        data: null
...
  parameters:
    X-localization:
      name: X-localization
      in: header
      description: 'Language (e.g. `en`, `ar`) (default will be: `en`)'
      schema:
        type: string
        enum:
          - en
          - ar
...
  securitySchemes:
    Bearer:
      type: apiKey
      name: Authorization
      in: header
      bearerFormat: JWT